Modifying the PUT and DELETE to use POST

+++ api/routes/genres.php
@@ -5,48 +5,51 @@ Flight::route('GET /genres', function(){
 	$result = Flight::get('database')->select("genres", "*");
 	
 	return Flight::json($result);
+
 });
 
 Flight::route('POST /genres', function(){
 	
 	$string = Flight::request()->body;
 	$genre = json_decode($string);
+	
 	$data = $genre->data;
+	$id = $genre->id;
+	$action = $genre->action;
 
-	$genreName = ucword($data);
+	if($action == 'POST'){
 
-	$result = Flight::get('database')->insert("genres", array('name' => $genreName));
+		$genreName = ucword($data);
 
-	$genre->id = $result;
-	$genre->name = $data;
+		$result = Flight::get('database')->insert("genres", array('name' => $genreName));
 
-	return Flight::json($genre);
-});
+		$genre->id = $result;
+		$genre->name = $data;
 
-Flight::route('DELETE /genres/@id', function($id){
+		return Flight::json($genre);
+
+	} else if($action == 'PUT'){
+
+		$name = $data->name;
+
+		$result = Flight::get('database')->update("genres", array('name' => $name), array('id' => $id));
+
+		return Flight::json($result);
+
+	} else if($action == 'DELETE') {
+		
+		$result = Flight::get('database')->delete("genres", array('id' => $id));
 	
-	$result = Flight::get('database')->delete("genres", array('id' => $id));
-	
-	return Flight::json($result);
+		return Flight::json($result);
+
+	}
+
 });
 
 Flight::route('GET /genres/@id', function($id){
 	
 	$result = Flight::get('database')->select("genres", "*", array('id' => $id));
 	
-	return Flight::json($result);
-});
-
-Flight::route('PUT /genres', function(){
-	$body = Flight::request()->body;
-
-	$data = json_decode($body)[0];
-	
-	$name = $data->name;
-	$id = $data->id;
-
-	$result = Flight::get('database')->update("genres", array('name' => $name), array('id' => $id));
-
 	return Flight::json($result);
 });
 
@@ -57,4 +60,5 @@ Flight::route('GET /genres/search', function(){
 	
 	return Flight::json($result);
 });
+
 ?>
